+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Examples.hs
@@ -129,7 +129,7 @@ mkLedgerExamples ::
   , PredicateFailure (EraRule "DELEGS" era) ~ ShelleyDelegsPredFailure era
   , PredicateFailure (EraRule "LEDGER" era) ~ ShelleyLedgerPredFailure era
   , Default (StashedAVVMAddresses era)
-  , ProtVerAtMost era 4
+  , AtMostEra "Mary" era
   ) =>
   (TxBody era -> [KeyPair 'Witness] -> TxWits era) ->
   Value era ->
@@ -358,9 +358,7 @@ exampleTxIns =
     [ TxIn (TxId (mkDummySafeHash @EraIndependentTxBody 1)) minBound
     ]
 
-exampleCerts ::
-  (ShelleyEraTxCert era, ProtVerAtMost era 8) =>
-  StrictSeq (TxCert era)
+exampleCerts :: (ShelleyEraTxCert era, AtMostEra "Babbage" era) => StrictSeq (TxCert era)
 exampleCerts =
   StrictSeq.fromList
     [ RegTxCert (keyToCredential exampleStakeKey)